Rename .report.computations → .report.operator
