+++ message_ix_models/report/operator.py
